+++ migrations/1.0.5.js
@@ -1,0 +1,21 @@
+'use strict';
+
+module.exports = async(db, redis) => {
+	console.log('Updating db for language settings, fixes broken 1.0.0 migration');
+	await db.collection('globalsettings').updateOne({ _id: 'globalsettings' }, {
+		'$set': {
+			'language': 'en-GB',
+			'boardDefaults.language': 'en-GB',
+		},
+	});
+	await db.collection('boards').updateMany({}, {
+		'$set': {
+			'settings.language': 'en-GB',
+		},
+	});	
+
+	console.log('Clearing globalsettings cache');
+	await redis.deletePattern('globalsettings');
+	console.log('Clearing boards cache');
+	await redis.deletePattern('board:*');
+};