+++ CHANGELOG.md
@@ -5,7 +5,9 @@
   - Make animated webp thumbnailing at least work and not give the above error (not supporting animated thumbs like .gif yet).
   - Add "ttf" extension to nginx snippets because some themes include fonts.
   - Reorganize <input> styling, use accent-color property on <input> elements, by @disco.
+  - Add globalmanage setting for board default language.
   - Fix an issue with the referer because of translation, and prevent invalid settings for it from breaking the web panel.
+  - Fix broken migration for board defaults causing language issues on newly created boards after updating from pre 1.0.0.
   - Npm audit.
 
 ### 1.0.4